+++ packages/moonstone/CHANGELOG.md
@@ -7,7 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Added
 
 - Support for high contrast colors in dark and light `moonstone`
-- `moonstone/BodyText` prop `nowrap` which automatically adds `moonstone/Marquee` support as well as limiting the content to only display one line of text
+- `moonstone/BodyText` prop `noWrap` which automatically adds `moonstone/Marquee` support as well as limiting the content to only display one line of text
 
 ### Fixed